Further rework of Dome logic
The operation of Dome light requires a combination of event TOGGLE_CABIN_LIGHTS (control 66579) and var LIGHT_POTENTIOMETER:7 (control 66911). The toggle should only be used to move switch between off and on. It should not be used between DIM and BRT. The frig in _BRT function requires a double press/call to reset Dome light to match local variable A32NX_Dome.